+++ sdk/python/capsem/models/container_spec.py
@@ -2,15 +2,16 @@
 
 from __future__ import annotations
 
-from pydantic import StrictStr
+from pydantic import StrictBool, StrictStr
 
 from .model_base import Model
 from .registry_access import RegistryAccess
 
 
 class ContainerSpec(Model):
-    nonnullable_optional = frozenset(['args', 'env'])
+    nonnullable_optional = frozenset(['args', 'attach', 'env'])
     args: list[StrictStr] | None = None
+    attach: StrictBool | None = None
     env: dict[str, StrictStr] | None = None
     image: StrictStr
     registry: RegistryAccess | None = None