+++ sdk/typescript/src/models/ContainerSpec.ts
@@ -4,6 +4,7 @@ import type { RegistryAccess } from "./RegistryAccess.js";
 
 export interface ContainerSpec {
   "args"?: Array<string>;
+  "attach"?: boolean;
   "env"?: Record<string, string>;
   "image": string;
   "registry"?: null | RegistryAccess;